dnsmasq: disable the service by default
The service will fail to start/stop when dnsmasq and
bind (provides named) are both installed since they
are using the same port 53:

| dnsmasq: failed to create listening socket for port 53: Address already in use

So disable the service by default, if the user really
want to use two dns servives, the config needs to be
changed to use different interfaces or ports.

As reference, other distro like centos/fedora also
disables the services by default on installation.

Signed-off-by: Jackie Huang <jackie.huang@windriver.com>
Signed-off-by: Jackie Huang <jackie.huang@windriver.com<mailto:jackie.huang@windriver.com>>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>